Update GetPortFromAddress test to compare against uint16.

+++ internal/socket/addr_port_test.go
@@ -8,11 +8,11 @@ import (
 )
 
 func Test_GetPortFromAddress(t *testing.T) {
-	require.Equal(t, 12345, GetPortFromAddress(&net.TCPAddr{
+	require.Equal(t, uint16(12345), GetPortFromAddress(&net.TCPAddr{
 		IP:   net.IPv4(1, 2, 3, 4),
 		Port: 12345,
 	}))
-	require.Equal(t, 12345, GetPortFromAddress(&net.UDPAddr{
+	require.Equal(t, uint16(12345), GetPortFromAddress(&net.UDPAddr{
 		IP:   net.IPv4(1, 2, 3, 4),
 		Port: 12345,
 	}))
